Fixed navbar and rearranged python pages
