+++ main.py
@@ -1,73 +1,58 @@
 ### IMPORTING PACKAGES ###
-from flask import Flask, render_template, request
+from flask import Flask, render_template, request, Blueprint
 from flask_login import login_required, current_user
-
+from __init__ import create_app
 import os
 
-app = Flask(__name__)
+main = Blueprint('main', __name__)
 
 
-
-@app.route("/")
+@main.route("/")
 def index():
     return render_template("auth/landingpage.html")
 
-@app.route("/landingpage/")
+@main.route("/landingpage/")
 def logout():
     return render_template("auth/landingpage.html")
 
-@app.route("/home/")
+@main.route("/home/")
 def home():
     return render_template('home.html')
 
-@app.route("/contact/")
+@main.route("/contact/")
 def contact():
     return render_template("Contact.html")
 
-@app.route("/Issues/")
+@main.route("/Issues/")
 def issues():
     return render_template("Issues.html")
 
-@app.route("/Maps/")
+@main.route("/Maps/")
 def maps():
     return render_template("Maps.html")
 
-@app.route("/subscription/")
+@main.route("/subscription/")
 def subscription():
     return render_template("SubscriptionPage.html")
 
-@app.route("/Timeline/")
+@main.route("/Timeline/")
 def timeline():
     return render_template("Timeline.html")
    
-@app.route("/worldsPage/")
+@main.route("/worldsPage/")
 def worlds():
     return render_template("worldsPage.html")
 
-@app.route("/worldinfo/")
+@main.route("/worldinfo/")
 def worldInfo():
     return render_template("worldinfo.html")
 
-@app.route("/userprofile/")
+@main.route("/userprofile/")
 def userprofile():
     return render_template("userProfile.html")
 
-####### Login code - WILL BE MOVED WHEN WORKING ON PROPER PYTHON FUNCTIONALITY - (-kye) ########
 
-database = {'Jane':'123',
-            'Dean':'666', 'Ollie':'999'}
-
-@app.route('/login', methods=['GET', 'POST']) # Defining the login page path
-def login(): # Log in page function
-    user = request.form['username']
-    pwd = request.form['password']
-    if user not in database:
-        return render_template('auth/landingpage.html', info='Invalid User!')
-    else:
-        if database[user] != pwd:
-            return render_template('auth/landingpage.html', info='Invalid Password!')
-        else:
-            return render_template('home.html', name=user)
+app = create_app()
 
 if __name__ == '__203.main__':
     app.run(debug=True)
