+++ static/login.css
@@ -1,0 +1,70 @@
+.field
+{
+    padding: 0.50em;
+}
+
+.toggle
+{
+    position: relative;
+    display: inline-block;
+    width: 3.75em;
+    height: 2em;
+    padding: 0;
+    
+}
+
+.toggle input
+{
+    opacity: 0;
+    width: 0;
+    height: 0;
+}
+
+.slider-round 
+{
+    position: absolute;
+    cursor: pointer;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: #ccc;
+    -webkit-transition: .4s;
+    transition: .4s;
+}
+
+.slider-round:before {
+    position: absolute;
+    content: "";
+    height: 1.62em;
+    width: 1.62em;
+    left: 0.25em;
+    bottom: 0.25em;
+    background-color: #ffffff;
+    -webkit-transition: .4s;
+    transition: .4s;
+  }
+  
+  input:checked + .slider-round {
+    background-color: #9dfddd;
+  }
+  
+  input:focus + .slider-round {
+    box-shadow: 0 0 1px #9dfddd;
+  }
+  
+  input:checked + .slider-round:before {
+    -webkit-transform: translateX(1.62em);
+    -ms-transform: translateX(1.62em);
+    transform: translateX(1.62em);
+  }
+  
+  /* Rounded sliders */
+  .slider-round {
+    padding: 0;
+    border-radius: 2.12em;
+  }
+  
+  .slider-round:before {
+    border-radius: 50%;
+  }
