+++ templates/religion.html
@@ -7,13 +7,13 @@
     </head>
     <body>
         <div class="navbar">
-            <a href="{{ url_for('home')}}">Home</a>
-            <a href="{{ url_for('worlds')}}">World Info</a>
-            <a href="{{ url_for('userprofile')}}">User Profile</a>
-            <a href="{{ url_for('issues')}}">Issues</a>
-            <a href="{{url_for('contact')}}">Contact</a>
-            <a href="{{ url_for('subscription')}}">Subscription</a>
-            <a href="{{ url_for('logout')}}">Log Out</a>
+            <a href="{{ url_for('main.home')}}">Home</a>
+            <a href="{{ url_for('main.worlds')}}">World Info</a>
+            <a href="{{ url_for('main.userprofile')}}">User Profile</a>
+            <a href="{{ url_for('main.issues')}}">Issues</a>
+            <a href="{{url_for('main.contact')}}">Contact</a>
+            <a href="{{ url_for('main.subscription')}}">Subscription</a>
+            <a href="{{ url_for('main.logout')}}">Log Out</a>
         </div>
     </body>
 </html>
